requirements.txt
flask
requests
psycopg2-binary
PyJWT
python-dotenv
